split trait metafunctions into bits
